requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
pydantic>=2.10.0
python-multipart>=0.0.12
pytest>=8.0.0
bcrypt==4.1.2
httpx>=0.27.0
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0